+++ pkg/webui/console/store/reducers/notifications.js
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { fillIntoArray, pageToIndices } from '@console/store/utils'
+
 import {
   GET_ARCHIVED_NOTIFICATIONS_SUCCESS,
   GET_INBOX_NOTIFICATIONS_SUCCESS,
@@ -26,22 +28,6 @@ const defaultState = {
     archived: { entities: [], totalCount: 0 },
   },
   unseenTotalCount: undefined,
-}
-
-// Update a range of values in an array by using another array and a start index.
-const fillIntoArray = (array, start, values, totalCount) => {
-  const newArray = [...array]
-  const end = Math.min(start + values.length, totalCount)
-  for (let i = start; i < end; i++) {
-    newArray[i] = values[i - start]
-  }
-  return newArray
-}
-
-const pageToIndices = (page, limit) => {
-  const startIndex = (page - 1) * limit
-  const stopIndex = page * limit - 1
-  return [startIndex, stopIndex]
 }
 
 const notifications = (state = defaultState, { type, payload }) => {
